msoffice/pptx/requirements.txt
python-pptx==0.6.23
Pillow>=10.0.0
lxml>=4.9.3
